add avatar to updateUser
