+++ helpers/index.js
@@ -2,10 +2,12 @@ const HttpError = require("./HttpError");
 const BadRequestError = require("./BadRequestError");
 const handleMongooseError = require("./handleMongooseError");
 const sendEmail = require("./sendEmail");
+const uploadToCloud = require("./uploadToCloud");
 
 module.exports = {
   HttpError,
   BadRequestError,
   handleMongooseError,
   sendEmail,
+  uploadToCloud
 };
