+++ middlewares/index.js
@@ -1,12 +1,11 @@
 const isValidId = require("./isValidId");
 const validateBody = require("./validateBody");
 const isAuthorized = require("./isAuthorized");
-const { uploadToCloud, upload } = require("./uploadToCloud");
+const upload = require("./upload");
 
 module.exports = {
   isValidId,
   validateBody,
   isAuthorized,
   upload,
-  uploadToCloud,
 };
